Made few changes/modifications summarised below:
Integrated the core game logic with the persistence entities.

Made the MemoryGameExceptions in the package
mjb.memorygame.game.exceptions checked exceptions.

Abstracted some entity / integration code to
mjb.memorygame.services.GameServiceImpl

Added lock to the signature of the
MemoryGame(List<Integer> cards, List<Integer> moves, boolean lock, int firstPlayer)
constructor. This is necessary because in order to reproduce
the required state of a game at the end of a turn the value
of lock tells us whether the completeTurn() method has been
called yet: if false completeTurn() must be called.